requirements.txt
python-telegram-bot==20.4
python-dotenv>=1.0.0
Flask>=2.0.0
psycopg2-binary
gspread
oauth2client
cryptography>=41.0.0